# openapi.yml
openapi: 3.0.3
info:
  title: AI Assistant API
  description: |
    Комплексная ИИ-платформа для семантического поиска и генерации RFC с пользовательскими настройками
  version: 2.0.0
  contact:
    name: AI Assistant Support
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/configurations/jira:
    post:
      summary: Create Jira configuration
      tags: [Data Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JiraConfigCreate'
      responses:
        '201':
          description: Конфигурация создана
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JiraConfigCreate:
      type: object
      properties:
        config_name:
          type: string
        jira_url:
          type: string
        username:
          type: string
        password:
          type: string
      required: [config_name, jira_url, username, password]
...